- synced with SVN

+++ Samples/CloudTest/CloudTest.cpp
@@ -7,7 +7,7 @@
  *  of patent rights can be found in the RakNet Patents.txt file in the same directory.
  *
  *
- *  Modified work: Copyright (c) 2016-2017, SLikeSoft UG (haftungsbeschränkt)
+ *  Modified work: Copyright (c) 2016-2018, SLikeSoft UG (haftungsbeschränkt)
  *
  *  This source code was modified by SLikeSoft. Modifications are licensed under the MIT-style
  *  license found in the license.txt file in the root directory of this source tree.
@@ -91,15 +91,15 @@ int main(void)
 		rakPeer[i]->Startup(RAKPEER_COUNT,&sd,1);
 	}
 
-	for (unsigned int i=SERVER_1; i < RAKPEER_COUNT; i++)
+	for (unsigned short i=SERVER_1; i < RAKPEER_COUNT; i++)
 	{
 		rakPeer[i]->SetMaximumIncomingConnections(RAKPEER_COUNT);
 	}
-	for (unsigned int i=CLIENT_1, j=CLOUD_CLIENT_1; i < SERVER_1; i++, j++)
+	for (unsigned short i=CLIENT_1, j=CLOUD_CLIENT_1; i < SERVER_1; i++, j++)
 	{
 		rakPeer[i]->AttachPlugin(&cloudClient[j]);
 	}
-	for (unsigned int i=SERVER_1, j=CLOUD_SERVER_1; i < RAKPEER_COUNT; i++, j++)
+	for (unsigned short i=SERVER_1, j=CLOUD_SERVER_1; i < RAKPEER_COUNT; i++, j++)
 	{
 		rakPeer[i]->AttachPlugin(&cloudServer[j]);
 	}
